Improve the code quality & refactoring
