+++ index.html
@@ -1,10 +1,11 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React</title>
+    <title>My Ecommerce</title>
+    <link rel="stylesheet" href="./assets/css/main.css" />
   </head>
   <body>
     <div id="root"></div>
