+++ public/assets/css/main.css
@@ -1,0 +1,148 @@
+.categoryList {
+  display: flex;
+  flex-direction: column;
+  width: 250px;
+  overflow-y: auto;
+  gap: 12px;
+  position: fixed;
+  top: 59px;
+  height: 100vh;
+  padding: 20px;
+}
+.productStyle {
+  width: 75%;
+  position: fixed;
+  right: 0;
+  overflow-y: auto;
+  height: 100vh;
+}
+
+.button {
+  display: inline-block;
+  padding: 12px 24px;
+  font-size: 16px;
+  font-weight: bold;
+  text-align: center;
+  text-transform: uppercase;
+  border: none;
+  border-radius: 5px;
+  background-color: #3498db;
+  color: white;
+  cursor: pointer;
+  transition: background-color 0.3s, transform 0.2s;
+}
+
+.button:hover {
+  background-color: #2980b9;
+  transform: scale(1.05);
+}
+.categories-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.category-link {
+  padding: 8px 16px;
+  font-size: 16px;
+  font-weight: bold;
+  border: 2px solid #3498db;
+  border-radius: 10px;
+  color: #3498db;
+  text-decoration: none;
+  margin: 0 10px;
+  transition: background-color 0.3s, color 0.3s;
+}
+
+.category-link:hover {
+  background-color: #3498db;
+  color: white;
+}
+
+/* Style for each product item */
+.product-container {
+  display: flex;
+  gap: 15px;
+  border: 4px solid black;
+  padding: 20px;
+  margin-bottom: 20px;
+  border-radius: 10px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+}
+
+.product-image {
+  width: 200px;
+  height: 150px;
+}
+
+.gradient-custom {
+  /* fallback for old browsers */
+  background: #6a11cb;
+
+  /* Chrome 10-25, Safari 5.1-6 */
+  background: -webkit-linear-gradient(
+    to right,
+    rgba(106, 17, 203, 1),
+    rgba(37, 117, 252, 1)
+  );
+
+  /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
+  background: linear-gradient(
+    to right,
+    rgba(106, 17, 203, 1),
+    rgba(37, 117, 252, 1)
+  );
+}
+/* Basic styling for the search box container */
+.searchBox {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: #f5f5f5;
+  padding: 10px;
+  border-radius: 8px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  margin: 0 auto;
+}
+
+/* Styling for the search input */
+.search {
+  border: none;
+  padding: 8px;
+  border-radius: 4px;
+  font-size: 14px;
+  width: 54%;
+}
+
+/* Styling for the search button */
+.searchButton {
+  background-color: #ff4500;
+  color: white;
+  padding: 8px 12px;
+  border-radius: 4px;
+  text-decoration: none;
+  font-size: 14px;
+  transition: background-color 0.3s ease-in-out;
+}
+
+/* Hover effect for the search button */
+.searchButton:hover {
+  background-color: #e63900;
+}
+.oneCard:hover {
+  transform: scale(1.03);
+  cursor: pointer;
+}
+
+/* Styling for the link */
+.titleLink {
+  color: #3498db;
+  text-decoration: none;
+  font-weight: bold;
+  transition: border-bottom 0.3s ease-in-out;
+}
+
+/* Hover effect */
+.titleLink:hover {
+  border-bottom: 2px solid #3498db;
+}
