--- src/DOCUMENTATION/doc.txt
@@ -1,33 +1,0 @@
-things i m gonna use    =>
-__________________________
-✨ REDUX=>
-✨ AXIOS=>for api calling
-✨ ERROR BOUNDARY=> (back to home)                                                           
-✨ CLASS COMPONET=> 4 error BOUNDARY
-✨ FUNCTION COMPONET =>
-✨ REACT ROUTER DOM=>
-✨ USEPARAMS =>
-✨ USEEFFECT =>
-✨ USEMEMO =>
-✨ pagination =>
-✨ infinite scroll=>
-uselocation
-
-
-
-
-thing i hv not used=>
-__________________________
-HOC => for handling loading and spinner
-context api
-usecallback 
-react.memo
-
-pges=>          (without css)
-__________________________
-✨main
-✨details
-✨category wise product page
-✨cart page
-
-not done => search product page
