+++ src/Features/SearchProduct/SearchProduct.Style.css
@@ -1,0 +1,78 @@
+/* Styling for the product list container */
+.product-list {
+  gap: 20px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+/* Styling for each product item */
+.product-item {
+  display: flex;
+  flex-direction: column;
+  background-color: white;
+  border-radius: 8px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  overflow: hidden;
+  transition: transform 0.2s ease-in-out;
+  width: 65%;
+}
+
+.product-item:hover {
+  transform: scale(1.03);
+}
+
+/* Styling for product image */
+.product-image img {
+  width: 100%;
+  height: auto;
+  object-fit: cover;
+  border-top-left-radius: 8px;
+  border-top-right-radius: 8px;
+}
+
+/* Styling for product details */
+.product-details {
+  padding: 20px;
+}
+
+/* Styling for product title */
+.product-title {
+  font-size: 18px;
+  color: #333;
+  text-decoration: none;
+  margin-bottom: 10px;
+  transition: color 0.2s ease-in-out;
+}
+
+.product-title:hover {
+  color: #d9534f;
+}
+
+/* Styling for product description */
+.product-description {
+  font-size: 14px;
+  color: #777;
+  margin-bottom: 15px;
+}
+
+/* Styling for cart button */
+.cart-button {
+  background-color: #d9534f;
+  color: white;
+  padding: 8px 16px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 14px;
+  transition: background-color 0.2s ease-in-out;
+}
+
+.cart-button.added {
+  background-color: #5bc0de;
+}
+
+.cart-button:hover {
+  background-color: #c9302c;
+}
